Fix launch security audit findings (CRITICAL/HIGH/MEDIUM/LOW)
Auth & disclosure:
- C1: stop leaking draft konflik to anonymous /cms/rest-map callers;
  filter drafts server-side in LocalServiceController index/kasusDetail,
  404 draft detail for non-owner/non-admin; remove unused /rest-map routes

XSS:
- C2: HTML-escape all DB strings in map.js + cms-konflik innerHTML sinks
  (stored XSS -> admin takeover via Livewire)

Seeder:
- C3: AdminUserSeeder uses forceCreate + random printed password
  (role/is_active/instansi were silently dropped by $fillable)

Privilege / file handling:
- H1: force status=draft for non-admins in TambahArtikel (Livewire tamper bypass)
- H2/H5/M3: lock down LFM — block rename-to-.php, use_package_routes=false,
  recurse blockPathTraversal over all inputs, alphanumeric filenames, 5MB cap
- H4: Livewire global upload mime allowlist, drop SVG preview (XSS)
- M1: rate-limit login (5/min) + single user lookup
- M4: whitelist role + admin re-check in EditUser
- M8: sanitize upload filenames (UUID+ext, drop client original name)

Hardening:
- L1: adminSession requires an active session (was anon-as-admin by accident)
- L2: standardize (int) session role casts
- L3: CmsUsers sort column allowlist

Tests: 88 passing (was 6 pre-existing failures on clean HEAD)
- fix strict in_array role whitelist that demoted admins (TambahUser/EditUser)
- fix logout test to POST (route is POST to prevent CSRF forced-logout)
- drop stale locale tests (locale routing disabled)
- add C1 regression test (anonymous must not see drafts)

.env untouched; production action items: APP_DEBUG=false, SESSION_SECURE_COOKIE
/SESSION_ENCRYPT=true, rotate external DB password, config:cache, nginx .php-deny
under /storage

+++ app/Livewire/EditArtikel.php
@@ -88,7 +88,7 @@ class EditArtikel extends Component
                     'updated_at' => now(),
                 ];
 
-                if (session('role_id') === 0) {
+                if ((int) session('role_id') === 0) {
                     $updateData['status'] = $this->status;
                 }
